Typescript refactor: Navbar.tsx
